Adding player, wall, and absolute map. Issues with player handling its position
Improving mobility of the player. Removing the player to be an extern variable. Overloading operators and adding some other goods.

Improving movement a little more. Now the it can go straight in every degree

Adding transformed map. Also, adding precompiled check.

Getting the original position of the player to use it as a reference

Added perspective transform. Moved the wall and the player a little more to the center.

Finishing the perspective transformation map