Translated the entire drawing algorithm, with success.
Incorrect buffer

It started to draw something, although it is not what I expected.

Fixing a couple of bugs. Now it is drawing the neighbors.

For now, I will push this, and I will use bisqwit code.

Found the bug with the equation

Adds the previous angle

Fixing issue with artifacts. The Y top array was initialized with garbage.